feat: Manage the errors of the Reservation entity

+++ src/main/java/com/grupo05/coworking_space/service/ReservationService.java
@@ -1,6 +1,7 @@
 package com.grupo05.coworking_space.service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
@@ -22,36 +23,69 @@ public class ReservationService {
     }
 
     public ReservationDTO createReservation(ReservationDTO reservationDTO) {
-        Reservation reservation = reservationMapper.convertToEntity(reservationDTO);
-        Reservation newReserve = reservationRepository.save(reservation);
-        return reservationMapper.convertToDTO(newReserve);
+        try {
+            if (reservationDTO == null) 
+                throw new RuntimeException("Los datos de la reserva no pueden ser nulos");
+            Reservation reservation = reservationMapper.convertToEntity(reservationDTO);
+            Reservation savedReservation = reservationRepository.save(reservation);
+            return reservationMapper.convertToDTO(savedReservation);
+        } catch (Exception e) {
+            throw new RuntimeException("Error en la creación de la reserva: " + e.getMessage());
+        }
     }
 
     public ReservationDTO findReservationByID(int id) {
-        Reservation reservation = reservationRepository.findById(id).get();
-        return reservationMapper.convertToDTO(reservation);
+        try{
+            Optional<Reservation> reservation = reservationRepository.findById(id);
+            if(!reservation.isPresent())
+                throw new RuntimeException("Reserva no encontrada");
+            return reservationMapper.convertToDTO(reservation.get());
+        } catch (Exception e) {
+            throw new RuntimeException("Error al buscar la reserva: " + e.getMessage());
+        }
     }
 
+    // findAll nunca devuelbe null
     public List<ReservationDTO> findAllReservations() {
-        List<Reservation> reservations = reservationRepository.findAll();
-        return reservations.stream()
-                .map(reservationMapper::convertToDTO)
-                .collect(Collectors.toList());
+        try{
+            List<Reservation> reservations = reservationRepository.findAll();
+            if (reservations.isEmpty()) 
+            throw new RuntimeException("No hay reservas en la base de datos");
+            return reservations.stream()
+            .map(reservationMapper::convertToDTO)
+            .collect(Collectors.toList());
+        } catch (Exception e) {
+            throw new RuntimeException("Error al obtener las reservas: " + e.getMessage());
+        }
     }
 
     public ReservationDTO updateResevation(ReservationDTO reservationDTO, int id) {
-        Reservation updateReservation = reservationRepository.findById(id).get();
-        updateReservation.setDateInit(reservationDTO.getDateInit());
-        updateReservation.setDateEnd(reservationDTO.getDateEnd());
-        updateReservation.setReserveStatus(reservationDTO.getReserveStatus());
-        updateReservation.setDescription(reservationDTO.getDescription());
-
-        Reservation savedReservation = reservationRepository.save(updateReservation);
-        return reservationMapper.convertToDTO(savedReservation);
+        try {
+            if (reservationDTO == null) 
+                throw new RuntimeException("Los datos de la reserva no pueden ser nulos");
+            Optional<Reservation> optionalReservation = reservationRepository.findById(id);
+            if(!optionalReservation.isPresent())
+                throw new RuntimeException("Reserva no encontrada");
+            Reservation updateReservation = optionalReservation.get();
+            updateReservation.setDateInit(reservationDTO.getDateInit());
+            updateReservation.setDateEnd(reservationDTO.getDateEnd());
+            updateReservation.setReserveStatus(reservationDTO.getReserveStatus());
+            updateReservation.setDescription(reservationDTO.getDescription());
+            Reservation savedReservation = reservationRepository.save(updateReservation);
+            return reservationMapper.convertToDTO(savedReservation);
+        } catch (Exception e) {
+            throw new RuntimeException("Error al actualizar la reserva: " + e.getMessage());
+        }
     }
 
     public void deleteReservation(int id) {
-        reservationRepository.deleteById(id);
+        try {
+            if (!reservationRepository.existsById(id)) 
+                throw new RuntimeException("No se puede eliminar. Reserva no encontrada con ID: " + id);
+            reservationRepository.deleteById(id);
+        } catch (Exception e) {
+            throw new RuntimeException("Error al eliminar la reserva: " + e.getMessage());
+        }
     }
 
 }
